20240119 ~ 20240131[은지/오류수정 및 기능 추가]

+++ src/main/java/com/cruise/project_cruise/controller/openbank/OpenBankController.java
@@ -1,16 +1,10 @@
 package com.cruise.project_cruise.controller.openbank;
 
 import com.cruise.project_cruise.dto.develop.OpenBankDTO;
-import com.cruise.project_cruise.service.CrewBoardService;
-import com.cruise.project_cruise.service.CrewCommentService;
-import com.cruise.project_cruise.service.DevelopOpenBankingService;
-import com.cruise.project_cruise.service.TemplateService;
+import com.cruise.project_cruise.service.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
-
-import javax.annotation.Resource;
-import java.util.ArrayList;
 import java.util.List;
 
 @RequestMapping(value="/develop/openbank")
@@ -19,15 +13,19 @@ public class OpenBankController {
 
     @Autowired
     private DevelopOpenBankingService developOpenBankingService;
+    @Autowired
+    private DevelopOpenBankUsingService developOpenBankUsingService;
 
     @RequestMapping(value="")
     public ModelAndView bankAccountList() throws Exception {
 
         ModelAndView mav = new ModelAndView();
+        String alertMsg = "";
 
         List<OpenBankDTO> accountList =
                 developOpenBankingService.getAccountList();
 
+        mav.addObject("alertMsg", alertMsg);
         mav.addObject("accountList",accountList);
         mav.setViewName("forDevelop/account");
 
@@ -37,6 +35,15 @@ public class OpenBankController {
     @PostMapping(value="/addAccount")
     public ModelAndView bankAccountAdd(OpenBankDTO openBankDTO) throws Exception {
         ModelAndView mav = new ModelAndView();
+
+        String alertMsg = "";
+
+        if(developOpenBankingService.getAccountList().contains(openBankDTO)) {
+            alertMsg = "등록하려는 계좌가 이미 등록되어 있습니다.";
+            mav.addObject("alertMsg",alertMsg);
+            mav.setViewName("forDevelop/account");
+            return mav;
+        }
 
         developOpenBankingService.insertAccount(openBankDTO);
         System.out.println("[Develop] 오픈뱅킹 테스트 계좌 insert : " + openBankDTO.getOpen_bank()+ " " +openBankDTO.getOpen_account());
@@ -57,8 +64,16 @@ public class OpenBankController {
     public ModelAndView bankAccountDelete(@RequestParam String account) throws Exception {
         ModelAndView mav = new ModelAndView();
 
+        if(!developOpenBankUsingService.getUsingList(account).isEmpty()) {
+            String alertMsg = "계좌내역이 존재하는 계좌는 삭제할 수 없습니다.";
+            mav.addObject("alertMsg",alertMsg);
+            mav.setViewName("forDevelop/account");
+            return mav;
+        }
+
         developOpenBankingService.deleteAccount(account);
         System.out.println("[Develop] 오픈뱅킹 테스트 계좌 Delete : " + account);
+
         mav.setViewName("redirect: ");
         return mav;
     }
